requirements.txt
streamlit
streamlit-authenticator
python-dotenv
requests
yfinance
pandas
plotly
feedparser
